check if Mender is enabled before initializing /etc/mender
Because:

* the startup script fails on a non-mender-image if it tries to
  initialize the mender configuration

This commit:

* adds a check if the file /etc/mender/mender.conf (added to the image
  by the mender-convert process) exists and only then initializes
  Mender

+++ armbian/base/scripts/systemd-startup-checks.sh
@@ -160,7 +160,7 @@ swapon /mnt/ssd/swapfile || true
 # Firmware updates
 # ------------------------------------------------------------------------------
 # initialize mender configuration
-if ! grep -q '/shift/' /etc/mender/mender.conf ; then
+if [[ -f /etc/mender/mender.conf ]] && ! grep -q '/shift/' /etc/mender/mender.conf ; then
     exec_overlayroot all-layers 'rm -f /etc/mender/mender.* /etc/mender/server.crt || true'
     exec_overlayroot all-layers 'cp -f /opt/shift/config/mender/mender.conf /etc/mender/'
 fi
